🚒 add passive mode test  * fix #40
