+++ config/default/agent_image_patch.yaml
@@ -8,5 +8,5 @@ spec:
     spec:
       containers:
       # Change the value of image field below to your controller image URL
-      - image: kubespheredev/porter-agent:cc1afae4
+      - image: kubespheredev/porter-agent:df370794
         name: porter-agent
