+++ config/manager/manager.yaml
@@ -48,7 +48,7 @@ spec:
       - command:
         - /manager
         image: controller:latest
-        imagePullPolicy: Always
+        imagePullPolicy: IfNotPresent
         name: manager
         readinessProbe:
           httpGet:
